requirements.txt
streamlit==1.28.0
pandas==2.1.0
plotly==5.18.0
yfinance==0.2.36
Pillow==10.0.0
requests==2.31.0
numpy==1.25.2
python-dotenv==1.0.0
kaleido  # For saving plotly figures as static images

